Rename cli file to app.go
